Add manual resume ability to wget boot.sh command
Add the OMARCHY_RESUME_INSTALL=true after the pipe before the bash and
it the boot.sh will skip removing the ~/.local/share/omarchy directory
and instead re-install with the current local repo.

This is helpful if you exited the installer to modify some code locally
after a failed install attempt, and you don't want to push a change to
github to retry the install.

It's similar to the gum prompt "Retry install" but supports that same
idea from the initial entry point of the wget command to boot.sh

Here's a full example:

wget -qO- https://raw.githubusercontent.com/basecamp/omarchy/master/boot.sh | OMARCHY_RESUME_INSTALL=true bash

+++ boot.sh
@@ -84,14 +84,21 @@ if [[ $EUID -eq 0 ]]; then
   exit 0 # exit to not run the rest of the script, and avoid cloning as root
 fi
 
-echo -e "\nCloning Omarchy from: https://github.com/${OMARCHY_REPO}.git"
-rm -rf ~/.local/share/omarchy/
-git clone "https://github.com/${OMARCHY_REPO}.git" ~/.local/share/omarchy >/dev/null
+if [[ -n $OMARCHY_RESUME_INSTALL ]]; then
+  echo -e "\n\e[32mResuming Omarchy installation from where it left off...\e[0m\n"
+else
+  echo -e "\nCloning Omarchy from: https://github.com/${OMARCHY_REPO}.git"
 
-echo -e "\e[32mUsing branch: $OMARCHY_REF\e[0m"
-cd ~/.local/share/omarchy
-git fetch origin "${OMARCHY_REF}" && git checkout "${OMARCHY_REF}"
-cd -
+  rm -rf ~/.local/share/omarchy/
+  git clone "https://github.com/${OMARCHY_REPO}.git" ~/.local/share/omarchy >/dev/null
+
+  if [[ $OMARCHY_REF != "master" ]]; then
+    echo -e "\e[32mUsing branch: $OMARCHY_REF\e[0m"
+    cd ~/.local/share/omarchy
+    git fetch origin "${OMARCHY_REF}" && git checkout "${OMARCHY_REF}"
+    cd -
+  fi
+fi
 
 echo -e "\nInstallation starting..."
 source ~/.local/share/omarchy/install.sh
